clean unused code
