--- packages/examples/fortune-v2/reputation-oracle/server/src/common/collection/index.ts
@@ -1,3 +1,0 @@
-export * from './pagination';
-export * from './search';
-export * from './sort';
