--- packages/examples/fortune-v2/reputation-oracle/server/src/common/decorators/base.ts
@@ -1,5 +1,0 @@
-export interface IBase {
-  id: number;
-  createdAt: Date;
-  updatedAt: Date;
-}
