+++ packages/examples/fortune-v2/reputation-oracle/server/src/common/decorators/index.ts
@@ -1,6 +1,1 @@
-export * from './user';
 export * from './public';
-export * from './role';
-export * from './job';
-export * from './webhook';
-export * from './reputation';
