--- packages/examples/fortune-v2/reputation-oracle/server/src/common/decorators/role.ts
@@ -1,6 +1,0 @@
-import { SetMetadata } from '@nestjs/common';
-
-export const Roles = <T>(
-  ...roles: Array<T>
-): ((target: any, key?: any, descriptor?: any) => any) =>
-  SetMetadata('roles', [...roles]);
