+++ packages/examples/fortune-v2/reputation-oracle/server/src/common/enums/collection.ts
@@ -1,0 +1,5 @@
+export enum SortDirection {
+    ASC = 'ASC',
+    DESC = 'DESC',
+  }
+
