+++ packages/examples/fortune-v2/reputation-oracle/server/src/common/enums/job.ts
@@ -1,19 +1,3 @@
-export enum JobStatus {
-  PENDING = 'PENDING',
-  PAID = 'PAID',
-  LAUNCHED = 'LAUNCHED',
-  COMPLETED = 'COMPLETED',
-  FAILED = 'FAILED',
-}
-
-export enum EscrowStatus {
-  LAUNCHED = 0,
-  PENDING = 1,
-  PAID = 2,
-  COMPLETE = 3,
-  CANCELLED = 4,
-}
-
 export enum JobMode {
   BATCH = 'BATCH',
   DESCRIPTIVE = 'DESCRIPTIVE',
