--- packages/examples/fortune-v2/reputation-oracle/server/src/common/enums/payment.ts
@@ -1,77 +1,0 @@
-export enum Currency {
-  USD = 'usd',
-  AED = 'aed',
-  ARS = 'ars',
-  AUD = 'aud',
-  BDT = 'bdt',
-  BMD = 'bmd',
-  BRL = 'brl',
-  CAD = 'cad',
-  CHF = 'chf',
-  CLP = 'clp',
-  CNY = 'cny',
-  CZK = 'czk',
-  DKK = 'dkk',
-  EUR = 'eur',
-  GBP = 'gbp',
-  HKD = 'hkd',
-  HUF = 'huf',
-  IDR = 'idr',
-  ILS = 'ils',
-  INR = 'inr',
-  JPY = 'jpy',
-  KRW = 'krw',
-  LKR = 'lkr',
-  MMK = 'mmk',
-  MXN = 'mxn',
-  MYR = 'myr',
-  NGN = 'ngn',
-  NOK = 'nok',
-  NZD = 'nzd',
-  PHP = 'php',
-  PKR = 'pkr',
-  PLN = 'pln',
-  RUB = 'rub',
-  SAR = 'sar',
-  SEK = 'sek',
-  SGD = 'sgd',
-  THB = 'thb',
-  TRY = 'try',
-  TWD = 'twd',
-  UAH = 'uah',
-  VND = 'vnd',
-  ZAR = 'zar',
-}
-
-export enum TokenName {
-  HMT = 'hmt',
-}
-
-export enum TokenId {
-  HUMAN_PROTOCOL = 'human-protocol',
-}
-
-export enum PaymentSource {
-  FIAT = 'fiat',
-  CRYPTO = 'crypto',
-  BALANCE = 'balance',
-}
-
-export enum PaymentFiatMethodType {
-  CARD = 'card',
-}
-
-export enum PaymentType {
-  DEPOSIT = 'DEPOSIT',
-  REFUND = 'REFUND',
-  WITHDRAWAL = 'WITHDRAWAL',
-}
-
-export enum PaymentStatus {
-  FAILED = 'FAILED',
-  SUCCEEDED = 'SUCCEEDED',
-}
-
-export enum CoingeckoId {
-  hmt = 'human-protocol',
-}
