+++ packages/examples/fortune-v2/reputation-oracle/server/src/common/enums/reputation.ts
@@ -1,0 +1,7 @@
+export enum ReputationEntityType {
+    WORKER = 'WORKER',
+    JOB_LAUNCHER = 'JOB_LAUNCHER',
+    EXCHANGE_ORACLE = 'EXCHANGE_ORACLE',
+    RECORDING_ORACLE = 'RECORDING_ORACLE',
+    REPUTATION_ORACLE = 'REPUTATION_ORACLE',
+  }
