--- packages/examples/fortune-v2/reputation-oracle/server/src/common/enums/user.ts
@@ -1,10 +1,0 @@
-export enum UserStatus {
-  ACTIVE = 'ACTIVE',
-  INACTIVE = 'INACTIVE',
-  PENDING = 'PENDING',
-}
-
-export enum UserType {
-  OPERATOR = 'OPERATOR',
-  REQUESTER = 'REQUESTER',
-}
