+++ packages/examples/fortune-v2/reputation-oracle/server/src/common/enums/webhook.ts
@@ -1,0 +1,6 @@
+export enum WebhookStatus {
+    PENDING = 'PENDING',
+    COMPLETED = 'COMPLETED',
+    FAILED = 'FAILED',
+    PAID = 'PAID',
+  }
