--- packages/examples/fortune-v2/reputation-oracle/server/src/common/test/constants.ts
@@ -1,25 +1,0 @@
-export const MOCK_REQUESTER_TITLE = 'Mock job title';
-export const MOCK_REQUESTER_DESCRIPTION = 'Mock job description';
-export const MOCK_FORTUNES_REQUIRED = 5;
-export const MOCK_CHAIN_ID = 1;
-export const MOCK_ADDRESS = '0x1234567890abcdef';
-export const MOCK_FILE_URL = 'mockedFileUrl';
-export const MOCK_FILE_HASH = 'mockedFileHash';
-export const MOCK_FILE_KEY = 'manifest.json';
-export const MOCK_PRIVATE_KEY =
-  'd334daf65a631f40549cc7de126d5a0016f32a2d00c49f94563f9737f7135e55';
-export const MOCK_BUCKET_NAME = 'bucket-name';
-export const MOCK_RECORDING_ORACLE_ADDRESS =
-  '0xCf88b3f1992458C2f5a229573c768D0E9F70C44e';
-export const MOCK_REPUTATION_ORACLE_ADDRESS =
-  '0x2E04d5D6cE3fF2261D0Cb04d41Fb4Cd67362A473';
-export const MOCK_EXCHANGE_ORACLE_WEBHOOK_URL = 'http://localhost:3000';
-export const MOCK_JOB_LAUNCHER_FEE = 5;
-export const MOCK_RECORDING_ORACLE_FEE = 5;
-export const MOCK_REPUTATION_ORACLE_FEE = 5;
-export const MOCK_TRANSACTION_HASH =
-  '0xd28e4c40571530afcb25ea1890e77b2d18c35f06049980ca4fb71829f64d89dc';
-export const MOCK_EMAIL = 'test@example.com';
-export const MOCK_CUSTOMER_ID = 'customer123';
-export const MOCK_PAYMENT_ID = 'payment123';
-export const MOCK_EXPIRES_IN = 1787260813820;
