+++ packages/examples/fortune-v2/reputation-oracle/server/src/modules/reputation/reputation.dto.ts
@@ -1,7 +1,7 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsEnum, IsNumber, IsString } from 'class-validator';
 import { ChainId } from '@human-protocol/sdk';
-import { ReputationEntityType } from '../../common/decorators';
+import { ReputationEntityType } from '../../common/enums';
 
 export class ReputationCreateDto {
   @ApiProperty()
