+++ packages/examples/fortune-v2/reputation-oracle/server/src/modules/reputation/reputation.service.spec.ts
@@ -1,10 +1,10 @@
 import { Test } from '@nestjs/testing';
 import { ReputationService } from './reputation.service';
 import { ReputationRepository } from './reputation.repository';
-import { ReputationEntityType } from '../../common/decorators';
+import { ReputationEntityType } from '../../common/enums';
 import { ChainId } from '@human-protocol/sdk';
 import { ReputationEntity } from './reputation.entity';
-import { MOCK_ADDRESS } from '../../common/test/constants';
+import { MOCK_ADDRESS } from '../../../test/constants';
 import { WebhookRepository } from '../webhook/webhook.repository';
 import { createMock } from '@golevelup/ts-jest';
