+++ packages/examples/fortune-v2/reputation-oracle/server/src/modules/webhook/webhook.cron.ts
@@ -3,9 +3,9 @@ import { WebhookService } from './webhook.service';
 import { Cron, CronExpression } from '@nestjs/schedule';
 import { WebhookRepository } from './webhook.repository';
 import { RETRIES_COUNT_THRESHOLD } from '../../common/constants';
-import { SortDirection } from '../../common/collection';
+import { SortDirection } from '../../common/enums';
 import { LessThanOrEqual } from 'typeorm';
-import { WebhookStatus } from '../../common/decorators';
+import { WebhookStatus } from '../../common/enums';
 
 @Injectable()
 export class WebhookCron {
