+++ packages/examples/fortune-v2/reputation-oracle/server/test/constants.ts
@@ -1,0 +1,9 @@
+export const MOCK_REQUESTER_TITLE = 'Mock job title';
+export const MOCK_REQUESTER_DESCRIPTION = 'Mock job description';
+export const MOCK_ADDRESS = '0x1234567890abcdef';
+export const MOCK_FILE_URL = 'mockedFileUrl';
+export const MOCK_FILE_HASH = 'mockedFileHash';
+export const MOCK_FILE_KEY = 'manifest.json';
+export const MOCK_JOB_LAUNCHER_FEE = 5;
+export const MOCK_RECORDING_ORACLE_FEE = 5;
+export const MOCK_REPUTATION_ORACLE_FEE = 5;
